move case display config to wrapper classes
