+++ corehq/ex-submodules/casexml/apps/case/templatetags/case_tags.py
@@ -24,10 +24,155 @@ register = template.Library()
 DYNAMIC_CASE_PROPERTIES_COLUMNS = 4
 
 
-def wrapped_case(case):
-    json = case.to_json()
-    case_class = CommCareCase.get_wrap_class(json)
-    return case_class.wrap(case.to_json())
+class CaseDisplayWrapper(object):
+    def __init__(self, case):
+        self.case = case
+        self.domain = case.domain
+        self.case_id = case.case_id
+
+    def actions(self):
+        actions = self.case.to_json()['actions']
+        return actions.reverse()
+
+    def to_full_dict(self):
+        """
+        Include calculated properties that need to be available to the case
+        details display by overriding this method.
+        """
+        json = self.case.to_json()
+        json['status'] = _('Closed') if self.case.closed else _('Open')
+
+        return json
+
+    def get_display_config(self):
+        return [
+            {
+                "layout": [
+                    [
+                        {
+                            "expr": "name",
+                            "name": _("Name"),
+                        },
+                        {
+                            "expr": "opened_on",
+                            "name": _("Opened On"),
+                            "parse_date": True,
+                            'is_phone_time': True,
+                        },
+                        {
+                            "expr": "modified_on",
+                            "name": _("Modified On"),
+                            "parse_date": True,
+                            "is_phone_time": True,
+                        },
+                        {
+                            "expr": "closed_on",
+                            "name": _("Closed On"),
+                            "parse_date": True,
+                            "is_phone_time": True,
+                        },
+                    ],
+                    [
+                        {
+                            "expr": "type",
+                            "name": _("Case Type"),
+                            "format": '<code>{0}</code>',
+                        },
+                        {
+                            "expr": "user_id",
+                            "name": _("Last Submitter"),
+                            "process": 'doc_info',
+                        },
+                        {
+                            "expr": "owner_id",
+                            "name": _("Owner"),
+                            "process": 'doc_info',
+                        },
+                        {
+                            "expr": "_id",
+                            "name": _("Case ID"),
+                        },
+                    ],
+                ],
+            }
+        ]
+
+    def dynamic_properties(self):
+        # pop seen properties off of remaining case properties
+        dynamic_data = self.case.dynamic_case_properties()
+        # hack - as of commcare 2.0, external id is basically a dynamic property
+        # so also check and add it here
+        if self.case.external_id:
+            dynamic_data['external_id'] = self.case.external_id
+
+        return dynamic_data
+
+
+class SupplyPointDisplayWrapper(CaseDisplayWrapper):
+    def to_full_dict(self):
+        from corehq.apps.locations.models import SQLLocation
+        data = super(SupplyPointDisplayWrapper, self).to_full_dict()
+        data.update({
+            'location_type': None,
+            'location_site_code': None,
+            'location_parent_name': None,
+        })
+        try:
+            location = SQLLocation.objects.get(location_id=self.case.location_id)
+        except SQLLocation.DoesNotExist:
+            pass
+        else:
+            data['location_type'] = location.location_type_name
+            data['location_site_code'] = location.site_code
+            if location.parent:
+                data['location_parent_name'] = location.parent.name
+
+        return data
+
+    def get_display_config(self):
+        return [
+            {
+                "layout": [
+                    [
+                        {
+                            "expr": "name",
+                            "name": _("Name"),
+                        },
+                        {
+                            "expr": "location_type",
+                            "name": _("Type"),
+                        },
+                        {
+                            "expr": "location_site_code",
+                            "name": _("Code"),
+                        },
+                        #{
+                            #"expr": "last_reported",
+                            #"name": _("Last Reported"),
+                        #},
+                    ],
+                    [
+                        {
+                            "expr": "location_parent_name",
+                            "name": _("Parent Location"),
+                        },
+                        {
+                            "expr": "owner_id",
+                            "name": _("Location"),
+                            "process": "doc_info",
+                        },
+                    ],
+                ],
+            }
+        ]
+
+
+def get_wrapped_case(case):
+    from corehq.apps.commtrack import const
+    wrapper_class = {
+        const.SUPPLY_POINT_CASE_TYPE: SupplyPointDisplayWrapper,
+    }.get(case.type, CaseDisplayWrapper)
+    return wrapper_class(case)
 
 
 def normalize_date(val):
@@ -45,24 +190,19 @@ def render_case(case, options):
     Change to kwargs when we're on a version of Django that does.
     """
     from corehq.apps.hqwebapp.templatetags.proptable_tags import get_tables_as_rows, get_default_definition
-    case = wrapped_case(case)
+    wrapped_case = get_wrapped_case(case)
     timezone = options.get('timezone', pytz.utc)
     timezone = timezone.localize(datetime.datetime.utcnow()).tzinfo
     _get_tables_as_rows = partial(get_tables_as_rows, timezone=timezone)
-    display = options.get('display') or case.get_display_config()
+    display = options.get('display') or wrapped_case.get_display_config()
     show_transaction_export = options.get('show_transaction_export') or False
     get_case_url = options['get_case_url']
 
-    data = copy.deepcopy(case.to_full_dict())
+    data = copy.deepcopy(wrapped_case.to_full_dict())
 
     default_properties = _get_tables_as_rows(data, display)
 
-    # pop seen properties off of remaining case properties
-    dynamic_data = case.dynamic_case_properties()
-    # hack - as of commcare 2.0, external id is basically a dynamic property
-    # so also check and add it here
-    if case.external_id:
-        dynamic_data['external_id'] = case.external_id
+    dynamic_data = wrapped_case.dynamic_properties()
 
     for section in display:
         for row in section['layout']:
@@ -78,9 +218,6 @@ def render_case(case, options):
     else:
         dynamic_properties = None
 
-    actions = case.to_json()['actions']
-    actions.reverse()
-
     the_time_is_now = datetime.datetime.utcnow()
     tz_offset_ms = int(timezone.utcoffset(the_time_is_now).total_seconds()) * 1000
     tz_abbrev = timezone.localize(the_time_is_now).tzname()
@@ -92,7 +229,7 @@ def render_case(case, options):
         except SQLProduct.DoesNotExist:
             return (_('Unknown Product ("{}")').format(product_id))
 
-    ledgers = get_current_ledger_transactions(case._id)
+    ledgers = get_current_ledger_transactions(wrapped_case.case_id)
     for section, product_map in ledgers.items():
         product_tuples = sorted(
             (_product_name(product_id), product_map[product_id]) for product_id in product_map
@@ -108,8 +245,8 @@ def render_case(case, options):
         "dynamic_properties_options": {
             "style": "table"
         },
-        "case": case,
-        "case_actions": mark_safe(json.dumps(actions)),
+        "case": wrapped_case.case,
+        "case_actions": mark_safe(json.dumps(wrapped_case.actions())),
         "timezone": timezone,
         "tz_abbrev": tz_abbrev,
         "case_hierarchy_options": {
@@ -120,7 +257,7 @@ def render_case(case, options):
         "ledgers": ledgers,
         "timezone_offset": tz_offset_ms,
         "show_transaction_export": show_transaction_export,
-        "xform_api_url": reverse('single_case_forms', args=[case.domain, case._id]),
+        "xform_api_url": reverse('single_case_forms', args=[wrapped_case.domain, wrapped_case.case_id]),
     })
 
 
@@ -344,7 +481,7 @@ def case_inline_display(case):
         if case.opened_on:
             ret = "%s (%s: %s)" % (case.name, _("Opened"), case.opened_on.date())
         else:
-            ret =  case.name
+            ret = case.name
     else:
         ret = _("Empty Case")
